refactor: update imports to use explicit module paths for UI components
