+++ src/app/admin/loading.tsx
@@ -1,4 +1,4 @@
-import { LoadingSpinner } from '@/components/ui'
+import { LoadingSpinner } from '@/components/ui/LoadingSpinner'
 
 export default function AdminLoading() {
   return (
